refactor: move utility functions to an utils file
