+++ src/deepmerge.ts
@@ -1,59 +1,21 @@
 import isPlainObj from "is-plain-obj"
 
+import {
+	cloneUnlessOtherwiseSpecified,
+	getKeys,
+	getMergeFunction,
+	propertyIsOnObject,
+	propertyIsUnsafe,
+} from "./utils"
+
 function defaultIsMergeable(value) {
 	return Array.isArray(value) || isPlainObj(value)
-}
-
-function emptyTarget(value) {
-	return Array.isArray(value) ? [] : {}
-}
-
-function cloneUnlessOtherwiseSpecified(value, options) {
-	return (options.clone && options.isMergeable(value))
-		? deepmerge(emptyTarget(value), value, options)
-		: value
 }
 
 function defaultArrayMerge(target, source, options) {
 	return target
 		.concat(source)
 		.map((element) => cloneUnlessOtherwiseSpecified(element, options))
-}
-
-function getMergeFunction(key, options) {
-	if (!options.customMerge) {
-		return deepmerge
-	}
-
-	const customMerge = options.customMerge(key)
-	return typeof customMerge === `function` ? customMerge : deepmerge
-}
-
-function getEnumerableOwnPropertySymbols(target) {
-	return Object.getOwnPropertySymbols
-		? Object.getOwnPropertySymbols(target).filter((symbol) =>
-			Object.prototype.propertyIsEnumerable.call(target, symbol),
-		)
-		: []
-}
-
-function getKeys(target) {
-	return Object.keys(target).concat(getEnumerableOwnPropertySymbols(target))
-}
-
-function propertyIsOnObject(object, property) {
-	try {
-		return property in object
-	} catch {
-		return false
-	}
-}
-
-// Protects from prototype poisoning and unexpected merging up the prototype chain.
-function propertyIsUnsafe(target, key) {
-	return propertyIsOnObject(target, key) // Properties are safe to merge if they don't exist in the target yet,
-		&& !(Object.hasOwnProperty.call(target, key) // unsafe if they exist up the prototype chain,
-			&& Object.propertyIsEnumerable.call(target, key)) // and also unsafe if they're nonenumerable.
 }
 
 function mergeObject(target, source, options) {
